refactor: flatten routes to combine influencer and brand dashboard
