--- app/routes/brand-home.tsx
@@ -1,31 +1,0 @@
-import type { LoaderFunction } from "@remix-run/node";
-import { redirect, json } from "@remix-run/node";
-import { useLoaderData } from "@remix-run/react";
-import { getUserSession } from "~/utils/userSession";
-import type { DbInfluencer } from "~/types/ApiOps";
-import { UserTypes } from "~/types/ApiOps";
-import { getUser } from "~/utils/db";
-import UserHomeCard from "~/components/UserHomeCard";
-
-export const loader: LoaderFunction = async (args) => {
-  const { userId, token } = await getUserSession(args);
-  if (!userId || !token) {
-    return redirect("/log-in");
-  }
-
-  const data = await getUser(token);
-  // TODO: change to DbBrand
-  const user = data.data as DbInfluencer;
-
-  if (user.usertype === UserTypes.INFLUENCER) {
-    return redirect("/influencer-home");
-  }
-
-  return json({ user, userId, token });
-};
-
-export default function BrandHome() {
-  const data = useLoaderData<any>();
-
-  return <UserHomeCard user={data.user} />;
-}
